# docs/swagger.yaml
definitions:
  dto.CartDTO:
    properties:
      products:
        items:
          $ref: '#/definitions/dto.CartItemDTO'
        type: array
      total_price:
        type: integer
    type: object
  dto.CartItemDTO:
    properties:
      cart_item_id:
        type: string
      category:
        type: string
      description:
        type: string
      name:
        type: string
      photo_url:
        type: string
      price:
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  dto.CreateCartItemDTO:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.CreateOrderCustomerDTO:
    properties:
      address:
        type: string
    required:
    - address
    type: object
...
  dto.OrderCustomerDTO:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      orderShopDTOs:
        items:
          $ref: '#/definitions/dto.OrderShopDTO'
        type: array
      payed:
        type: string
      total_price:
        type: integer
    required:
    - address
    - created_at
    - id
    - orderShopDTOs
    - payed
    - total_price
    type: object
  dto.OrderShopDTO:
    properties:
      id:
        type: string
      order_shop_items:
        items:
          $ref: '#/definitions/dto.OrderShopItemDTO'
        type: array
      shop_id:
        type: string
      status:
        type: string
    required:
    - id
    - order_shop_items
    - shop_id
    - status
    type: object
  dto.OrderShopItemDTO:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      photo_url:
        type: string
      price:
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - category
    - description
    - name
    - photo_url
    - price
    - product_id
    - quantity
    type: object
...
  dto.RefreshDTO:
    properties:
      fingerprint:
        type: string
    required:
    - fingerprint
    type: object
...
  dto.SignInDTO:
    properties:
      email:
        type: string
      fingerprint:
        type: string
      password:
        type: string
    required:
    - email
    - fingerprint
    - password
    type: object
  dto.SignUpDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: integer
      surname:
        type: string
    required:
    - email
    - name
    - password
    - surname
    type: object
  dto.UpdateCartItemDTO:
    properties:
      cart_product_id:
        type: string
      quantity:
        type: integer
    required:
    - cart_product_id
    - quantity
    type: object
...
info:
  contact:
    email: emir2701@yandex.ru
    name: API Support
    url: https://t.me/Emir_Shimshir
  description: This is simple api for marketplace
  title: Marketplace API
  version: "1.0"
paths:
  /api/v1/auth/logout:
    post:
      description: logout
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      description: refresh
      parameters:
      - description: fingerprint
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshDTO'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Refresh
      tags:
      - auth
  /api/v1/auth/sign-in:
    post:
      description: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignInDTO'
      responses: {}
      summary: SignIn
      tags:
      - auth
  /api/v1/auth/sign-up:
    post:
      description: create account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpDTO'
      responses: {}
      summary: SignUp
      tags:
      - auth
  /api/v1/cart/customer:
    get:
      description: get cart
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartDTO'
      security:
      - ApiKeyAuth: []
      summary: getCart
      tags:
      - cart
    post:
      description: create cart product
      parameters:
      - description: cart product info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCartItemDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartItemDTO'
      security:
      - ApiKeyAuth: []
      summary: createCartItem
      tags:
      - cart
    put:
      description: update cart product
      parameters:
      - description: cart product info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCartItemDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartItemDTO'
      security:
      - ApiKeyAuth: []
      summary: updateCartItem
      tags:
      - cart
  /api/v1/cart/customer/{cart_product_id}:
    delete:
      description: delete cart product
      parameters:
      - description: cartItemID
        in: path
        name: cart_product_id
        required: true
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: deleteCartItem
      tags:
      - cart
  /api/v1/order/customer:
    get:
      description: get order customer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OrderCustomerDTO'
            type: array
      security:
      - ApiKeyAuth: []
      summary: getOrderCustomersByCustomerID
      tags:
      - order
    post:
      description: create order customer
      parameters:
      - description: order customer info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderCustomerDTO'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: createOrderCustomer
      tags:
      - order
  /api/v1/order/seller/{order_shop_id}:
    put:
      description: update order shop status
      parameters:
      - description: orderShopID
        in: path
        name: order_shop_id
        required: true
        type: string
      - description: status
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OrderShopDTO'
            type: array
      security:
      - ApiKeyAuth: []
      summary: updateOrderShopStatusByShopID
      tags:
      - order
  /api/v1/order/seller/{shop_id}:
    get:
      description: get order shops
      parameters:
      - description: shopID
        in: path
        name: shop_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OrderShopDTO'
            type: array
      security:
      - ApiKeyAuth: []
      summary: getOrderShopsByShopID
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"